requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pytest>=8.0.0
pytest-playwright>=0.5.0
requests>=2.31.0
